Propagate QImage with file path in text to PointCloudCreator
Instead of propagating the selection of indices to PointCloudCreator,
propagate QImages with the path put in text field "path" of the QImage.
This way PointCloudCreator is unaware of _how_ the images are selected and
where they come from, it just knows _which_ images to use and some
human readable information for display.